Header/source refactor
