+++ src/controllers/filters/parser/Tokenizer.hpp
@@ -29,150 +29,29 @@ static const QRegularExpression tokenRegex(
 class Tokenizer
 {
 public:
-    Tokenizer(const QString &text)
-    {
-        QRegularExpressionMatchIterator i = tokenRegex.globalMatch(text);
-        while (i.hasNext())
-        {
-            auto text = i.next().captured();
-            this->tokens_ << text;
-            this->tokenTypes_ << tokenize(text);
-        }
-    }
+    Tokenizer(const QString &text);
 
-    bool hasNext() const
-    {
-        return this->i_ < this->tokens_.length();
-    }
+    bool hasNext() const;
+    QString next();
+    TokenType nextTokenType() const;
+    TokenType tokenType() const;
 
-    QString next()
-    {
-        this->i_++;
-        return this->tokens_.at(this->i_ - 1);
-    }
+    bool nextTokenIsBinaryOp() const;
+    bool nextTokenIsUnaryOp() const;
+    bool nextTokenIsMathOp() const;
 
-    TokenType nextTokenType() const
-    {
-        return this->tokenTypes_.at(this->i_);
-    }
+    void debug();
+    const QStringList allTokens();
 
-    TokenType tokenType() const
-    {
-        return this->tokenTypes_.at(this->i_ - 1);
-    }
-
-    bool nextTokenIsBinaryOp() const
-    {
-        return this->typeIsBinaryOp(this->nextTokenType());
-    }
-
-    bool typeIsBinaryOp(TokenType token) const
-    {
-        return token > TokenType::BINARY_START && token < TokenType::BINARY_END;
-    }
-
-    bool nextTokenIsUnaryOp() const
-    {
-        return this->typeIsUnaryOp(this->nextTokenType());
-    }
-
-    bool typeIsUnaryOp(TokenType token) const
-    {
-        return token > TokenType::UNARY_START && token < TokenType::UNARY_END;
-    }
-
-    bool nextTokenIsMathOp() const
-    {
-        return this->typeIsMathOp(this->nextTokenType());
-    }
-
-    bool typeIsMathOp(TokenType token) const
-    {
-        return token > TokenType::MATH_START && token < TokenType::MATH_END;
-    }
-
-    void debug()
-    {
-        if (this->i_ > 0)
-        {
-            qDebug() << "= current" << this->tokens_.at(this->i_ - 1);
-            qDebug() << "= current type" << this->tokenTypes_.at(this->i_ - 1);
-        }
-        else
-        {
-            qDebug() << "= no current";
-        }
-        if (this->hasNext())
-        {
-            qDebug() << "= next" << this->tokens_.at(this->i_);
-            qDebug() << "= next type" << this->tokenTypes_.at(this->i_);
-        }
-        else
-        {
-            qDebug() << "= no next";
-        }
-    }
-
-    const QStringList allTokens()
-    {
-        return this->tokens_;
-    }
+    static bool typeIsBinaryOp(TokenType token);
+    static bool typeIsUnaryOp(TokenType token);
+    static bool typeIsMathOp(TokenType token);
 
 private:
     int i_ = 0;
     QStringList tokens_;
     QList<TokenType> tokenTypes_;
 
-    TokenType tokenize(const QString &text)
-    {
-        if (text == "&&")
-            return TokenType::AND;
-        else if (text == "||")
-            return TokenType::OR;
-        else if (text == "(")
-            return TokenType::LP;
-        else if (text == ")")
-            return TokenType::RP;
-        else if (text == "+")
-            return TokenType::PLUS;
-        else if (text == "-")
-            return TokenType::MINUS;
-        else if (text == "*")
-            return TokenType::MULTIPLY;
-        else if (text == "/")
-            return TokenType::DIVIDE;
-        else if (text == "==")
-            return TokenType::EQ;
-        else if (text == "!=")
-            return TokenType::NEQ;
-        else if (text == "%")
-            return TokenType::MOD;
-        else if (text == "<")
-            return TokenType::LT;
-        else if (text == ">")
-            return TokenType::GT;
-        else if (text == "<=")
-            return TokenType::LTE;
-        else if (text == ">=")
-            return TokenType::GTE;
-        else if (text == "contains")
-            return TokenType::CONTAINS;
-        else if (text == "!")
-            return TokenType::NOT;
-        else
-        {
-            if (text.front() == '"' && text.back() == '"')
-                return TokenType::STRING;
-
-            if (validIdentifiers.contains(text))
-                return TokenType::IDENTIFIER;
-
-            bool flag;
-            if (text.toInt(&flag); flag)
-                return TokenType::INT;
-        }
-
-        return TokenType::NONE;
-    }
+    TokenType tokenize(const QString &text);
 };
 }  // namespace filterparser
